pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "iqa-async-microservice"
version = "0.1.0"
description = "Async Image Quality Assessment microservice"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.0",
    "pillow>=10.0.0",
    "httpx>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "hypothesis>=6.80.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.24.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "adapters*", "api*", "worker*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
